more sketching
Explored exhaustive /* routes based on children. It can be done. Most
likely candidate is making `extractPath` return an array of all possible
child routes (separated with /*). That requires a change in matching
where /*abcdefg is treated differently from /*. We probably also need
to change scoring to include length for equal scores (or use length as
a base of the score).